backend/requirements.txt
faiss-cpu==1.9.0.post1
fastapi==0.115.6
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
openai==1.59.7
pydantic==2.10.5
python-dotenv==1.0.1
uvicorn==0.34.0
